Split out morton number
